Fix for numerous bugs/problems identified by Jean-Luc Fattebert at LLNL -- Thank You!.
-- Removed adrivers.c and mdrivers.c from external_src/oomph_parmetis_3.1.1.
   They're not needed and cause problems for some linkers.

-- Replaced oomph_metis_4.0 by oomph_metis_from_parmetis_3.1.1
   (the serial version of metis that is shipped with (and therefore
   fully compatible with) parmetis_3.1.1.

-- New metis revealed a few mpi driver codes that didn't read in the
   partition from disk when doing self test. Fixed these.

-- Changed a catch by copy to catch by reference in mesh smoother.

-- Removed unnecessary includes in demo code that illustrates problem
   with missing "this->"

-- Removed unnecesary template<> in various axisymmetric multi-physics
   files. 

-- Forced an explicit instantiation of the HelmholtzEquations in 1,2
   and 3 dimensions.

-- Added compute_exact_Z2_error(...) to PseudoSolidNodeUpdateElement and
   its refineable counterpart (it forwards this to the BASIC element;
   a bit naughty -- won't work if the BASIC element isn't an 
   ElementWithZ2ErrorEstimator...).

-- Tidied up a few conversion warnings (double to int etc)

-- Added jeffrey orbit directory to subdir list in navier stokes doc 
   directory.


git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@241 595e1d82-adf1-4282-bac7-b6073fa2ca8e

+++ demo_drivers/interaction/unsteady_vmtk_fsi/unsteady_vmtk_fsi.cc
@@ -1295,7 +1295,7 @@ int main(int argc, char **argv)
  doc_info.number()++;   
 
  //Set value of dt
- double nsteps_per_period=40;
+ unsigned nsteps_per_period=40;
  double dt = Global_Parameters::Period/double(nsteps_per_period);
  problem.initialise_dt(dt);
